requirements.txt
appdirs==1.4.4
black==21.6b0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
pdf2image==1.15.1
Pillow==8.2.0
pytesseract==0.3.7
regex==2021.4.4
toml==0.10.2
Werkzeug==2.0.1
